EWE: Fixed issue related to switching mode from Staging to Production in EWE automatically deletes staging record. (EW-88)

+++ EIWS_BLL_Core/FormSetting.cs
@@ -113,8 +113,10 @@ namespace Epi.Web.BLL
                 this.FormSettingDao.UpDateSettingsList(FormSettingBO, FormSettingDTO.FormId);
 
                 // Clear all Draft records
+                if (FormSettingDTO.DeleteDraftData)
+                { 
                 this.FormSettingDao.DeleteDraftRecords(FormSettingDTO.FormId);
-
+                }
 
                List<UserBO> AdminList =  this.UserDao.GetAdminsBySelectedOrgs(FormSettingBO, FormSettingDTO.FormId);
